[IMP] a Reference holds now any sort of classes, no longer limited to 'binders', 'mappers', 'synchronizers' and 'backend adapters'
To obtain the class to use, one should now give the base class (Binder, Mapper,
...).  This will allow to add more types of classes to a Reference without
changing its code.  This will also allow to ask more precise classes, ie. we
can subclass 'Synchronizer' with 'ExportSynchronizer' and 'ImportSynchronizer'
and when we search the class to use, we ask for a subclass of
'ExportSynchronizer'.