[IMP] Allow configuring runner from config file
In addition to the ODOO_CONNECTOR_CHANNELS environment variables, we can
configure the runner channels in the odoo configuration file like this:

    [options]
    connector_channels = root:4

As such we can configure the job runner entirely with config files and
command line arguments.

We still need to load the `connector` module on the command line, as
there is no way yet pass the `load` parameter through the config file.