Add a new method to refine match of components
This method can be used to filter out some components. It is executed on
every candidate component returned by a lookup for a
collection/usage/model so must be kept light to execute if possible.

A use case can be for instance to have more than one kind of export in a
connector for a model.

Say that you have a different way to export customers and suppliers:

    with self.work_on('res.partner', kind='customer') as work:
        exporter = work.component(usage='record.exporter)
        exporter.run()

    with self.work_on('res.partner', kind='supplier') as work:
        exporter = work.component(usage='record.exporter)
        exporter.run()

You can declare your components as:

  class CustomerExporter(Component):

      _name = 'customer.exporter'
      _inherit = ['base.exporter']
      _usage = 'record.exporter'
      _apply_on = ['res.partner']

      @classmethod
      def _component_match(cls, work):
          return work.kind == 'customer'

      def run(self):
          # export customer

  class SupplierExporter(Component):

      _name = 'supplier.exporter'
      _inherit = ['base.exporter']
      _usage = 'record.exporter'
      _apply_on = ['res.partner']

      @classmethod
      def _component_match(cls, work):
          return work.kind == 'supplier'

      def run(self):
          # export supplier