Exclude current addon in ComponentRegistryCase setup
When we run a ComponentRegistryCase test using odoo's way to run tests
(--test-enable), the current addon is in state 'to install' or 'to
upgrade', thus its components not loaded (which is what we want in this
test Case, as it allow us to load manually the components / load
different components). When we run the same test Case from pytest or
nosetest, which are not run during an odoo upgrade, the state of the
addon is 'installed', thus components are loaded.

To ensure the same behavior with --test-enable and external tests
runners, the current addon is always excluded from the components to
load.