Refine lookup on specific model over generic component
With components having a generic implementation and another
one linked with a model such as:

```python

  class GenericExporter(Component):

      _name = 'generic.exporter'
      _inherit = ['base.exporter']
      _usage = 'record.exporter'

      def run(self, records):
          return self._export_items(records)

  class SaleExporter(Component):

      _name = 'sale.exporter'
      _inherit = ['generic.exporter']
      _apply_on = 'sale.order'

      def run(self, records):
          return self.do_stuff(records)
```

The previous behavior with this code:

```python
exporter = work.component('record.exporter')
```

Was to raise an error:
```
SeveralComponentError: Several components found for collection 'xx.backend',
usage 'record.exporter', model_name 'sale.order'. Found: [<class
'odoo.addons.component.core.generic.exporter'>, <class
'odoo.addons.component.core.sale.exporter'>]
```

This commit refines the lookup so if a specific Component is linked with the
asked model, it will be returned over the generic one. It allows to build
generic components and override them only in some use cases.

Follows the motivation of #272 but as a model's level instead of collection.

Closes #276