Exclude components created by tests from the list of addon's components
If components are declared in tests, exclude them from the "components
of the addon" list. If not, when we use the "load_components" method,
all the test components would be loaded.  This should never be an issue
when running the app normally, as the Python tests should never be
executed. But this is an issue when a test creates a test components for
the purpose of the test, then a second tests uses the "load_components"
to load all the addons of the module: it will load the component of the
previous test.