[IMP] component_event: do not raise events during install through UI
During install / upgrade of addons, the odoo registry is recreated and not
flagged as ready. At the end of the odoo registry loading, the registry
of components will be built. If we install component_event from the UI
(not with -i), the component's registry is already flagged as "ready",
but the 'base.event.collecter' component is not yet referenced in it.