component: allow ctx override at init
Before this change, the only way to set something in the ctx is to override the env property.
This means that the code would be executed N times and a new env initialized N times.

The original env is preserved and in any case is available on work.env and collection.env.